feat: add shared health starter
