+++ shared-lib/shared-starters/starter-health/.gitignore
@@ -1,0 +1,1 @@
+target/
