+++ shared-lib/shared-starters/starter-health/README.md
@@ -1,0 +1,3 @@
+# Shared Starter - Health
+
+Provides generic health endpoints for LMS services.
