+++ shared-lib/shared-starters/starter-health/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+com.shared.starter_health.HealthAutoConfiguration
